Make some modifications off of Dr. Qi Chu's feedback
This doesn't quite cover all of the feedback, which also suggested some
changes for replacing the "must be used for synchronization" stuff. I'm
still trying to figure out how to word it, but the suggestion can be
found below:

The issue is also for single detector triggering. This is presenting an
issue when only one detector is in "science mode" whilst the others are
temporarily offline, which is currently happening roughly 30% of the
time in O3. At the moment, SPIIR is not able to trigger on single
detectors, and therefore needs to implement some strategies in order to
deal with the scenario. Some options include ignoring less sensitive
detectors for coherent processing, but using them for SNR calculations.